+++ visualization/app/codeCharta/services/ajs-upgraded-providers.ts
@@ -1,13 +1,19 @@
 import { InjectionToken, Injector } from "@angular/core"
 
 export const ThreeSceneServiceToken = new InjectionToken("ThreeSceneService")
-
-export function ThreeSceneServiceFactory(injector: Injector) {
-	return injector.get("threeSceneService")
-}
-
 export const threeSceneServiceProvider = {
 	provide: ThreeSceneServiceToken,
-	useFactory: ThreeSceneServiceFactory,
+	useFactory: function ThreeSceneServiceFactory(injector: Injector) {
+		return injector.get("threeSceneService")
+	},
+	deps: ["$injector"]
+}
+
+export const CodeChartaServiceToken = new InjectionToken("CodeChartaService")
+export const codeChartaServiceProvider = {
+	provide: CodeChartaServiceToken,
+	useFactory: function CodeChartaServiceFactory(injector: Injector) {
+		return injector.get("codeChartaService")
+	},
 	deps: ["$injector"]
 }