+++ visualization/app/codeCharta/ui/customConfigs/customConfigs.module.ts
@@ -4,9 +4,11 @@ import { customConfigsComponent } from "./customConfigs.component"
 import { AddCustomConfigButtonComponent } from "./addCustomConfigButton/addCustomConfigButton.component"
 import { downgradeComponent } from "@angular/upgrade/static"
 import { DownloadCustomConfigsButtonComponent } from "./downloadCustomConfigsButton/downloadCustomConfigsButton.component"
+import { UploadCustomConfigButtonComponent } from "./uploadCustomConfigButton/uploadCustomConfigButton.component"
 
 angular
 	.module("app.codeCharta.ui.customConfigs", ["app.codeCharta.state"])
 	.component(customConfigsComponent.selector, customConfigsComponent)
 	.directive("ccAddCustomConfigButton", downgradeComponent({ component: AddCustomConfigButtonComponent }))
 	.directive("ccDownloadCustomConfigsButton", downgradeComponent({ component: DownloadCustomConfigsButtonComponent }))
+	.directive("ccUploadCustomConfigButton", downgradeComponent({ component: UploadCustomConfigButtonComponent }))